+++ presidio_evaluator/data_generator/presidio_pseudonymize.py
@@ -3,7 +3,7 @@ from typing import List
 from presidio_analyzer import RecognizerResult
 from presidio_anonymizer import AnonymizerEngine
 
-from presidio_evaluator.data_generator import SentenceFaker
+from presidio_evaluator.data_generator.faker_extensions.sentences import SentenceFaker
 
 
 class PresidioPseudonymization(SentenceFaker):
